deepdb/requirements.txt
numpy
tqdm
pandas
scikit-learn
git+https://github.com/SPFlow/SPFlow.git
pickle5
sqlparse
networkx
tables
duckdb==0.10.1
psycopg2-binary
func_timeout
bloom-filter
mpmath
statsmodels
scipy==1.11